Add files via upload
This document is a complimentary material for AI courses in Said Business School. It gives a high-level overview of the main AI terms that you come across in the AI course and serves as starting point for further research. We provide  more technical details on the topics, which are relevant to the course, while skipping very basic ones and more difficult subjects.  All the references are available in the text via hyperlinks.